Catch resource existance exceptions when resource isn't found
If the resource doesn't exist in the solr datastore when we're looking
for solr datastore related information we should return an empty list
of geospatial fields not raise an exception. The exception itself
comes from the datastore search action, not the solr one, because if
the resource isn't in the solr datastore then we ignore it and pass the
action on to the standard datastore itself.